backend-microservices/auth_microservices/requirements.txt

Flask
Werkzeug
Flask-SQLAlchemy
Flask-Cors
PyJWT